+++ homematicip/home.py
@@ -1,6 +1,8 @@
 import logging
 import threading
 import websocket
+import ssl
+import sys
 from typing import List
 
 from homematicip.EventHook import *
@@ -154,7 +156,8 @@ class Home(HomeMaticIPObject):
         self.apExchangeState = ApExchangeState.NONE
         self.availableAPVersion = None
         self.carrierSense = None
-        #:bool:displays if the access point is connected to the hmip cloud or not
+        #:bool:displays if the access point is connected to the hmip cloud or
+        # not
         self.connected = None
         #:str:the current version of the access point
         self.currentAPVersion = None
@@ -178,6 +181,8 @@ class Home(HomeMaticIPObject):
         self.__webSocketThread = None
         self.onEvent = EventHook()
         self.onWsError = EventHook()
+        #:bool:switch to enable/disable automatic reconnection of the websocket (default=True)
+        self.websocket_reconnect_on_error = True
 
         #:List[Device]: a collection of all devices in home
         self.devices = []
@@ -702,6 +707,7 @@ class Home(HomeMaticIPObject):
 
     def enable_events(self):
         websocket.enableTrace(True)
+
         self.__webSocket = websocket.WebSocketApp(
             self._connection.urlWebSocket,
             header=[
@@ -710,17 +716,37 @@ class Home(HomeMaticIPObject):
             ],
             on_message=self._ws_on_message,
             on_error=self._ws_on_error,
+            on_close=self._ws_on_close,
         )
-        self.__webSocketThread = threading.Thread(target=self.__webSocket.run_forever)
-        self.__webSocketThread.daemon = True
+
+        websocket_kwargs = {"ping_interval": 3}
+        if hasattr(sys, "_called_from_test"):  # disable ssl during a test run
+            sslopt = {"cert_reqs": ssl.CERT_NONE}
+            websocket_kwargs = {"sslopt": sslopt, "ping_interval": 2, "ping_timeout": 1}
+
+        self.__webSocketThread = threading.Thread(
+            name="hmip-websocket",
+            target=self.__webSocket.run_forever,
+            kwargs=websocket_kwargs,
+        )
+        self.__webSocketThread.setDaemon(True)
         self.__webSocketThread.start()
 
     def disable_events(self):
-        self.__webSocket.close()
+        if self.__webSocket:
+            self.__webSocket.close()
+            self.__webSocket = None
+
+    def _ws_on_close(self):
+        self.__webSocket = None
 
     def _ws_on_error(self, err):
         LOGGER.exception(err)
         self.onWsError.fire(err)
+        if self.websocket_reconnect_on_error:
+            logger.debug("Trying to reconnect websocket")
+            self.disable_events()
+            self.enable_events()
 
     def _ws_on_message(self, message):
         # json.loads doesn't support bytes as parameter before python 3.6
@@ -798,11 +824,11 @@ class Home(HomeMaticIPObject):
 
                 # TODO: implement INCLUSION_REQUESTED, NONE
                 eventList.append({"eventType": pushEventType, "data": obj})
-            except ValueError as valerr:
+            except ValueError as valerr:  # pragma: no cover
                 LOGGER.warning(
                     "Uknown EventType '%s' Data: %s", event["pushEventType"], event
                 )
 
-            except Exception as err:
+            except Exception as err:  # pragma: no cover
                 LOGGER.exception(err)
         self.onEvent.fire(eventList)