+++ homematicip_demo/fake_cloud_server.py
@@ -1,12 +1,15 @@
 import hashlib
 import json
 import time
+import asyncio
+import functools
+from aiohttp import web
 from datetime import datetime, timedelta
 from pathlib import Path
 from werkzeug.wrappers import Request, Response
 
 
-class FakeCloudServer:
+class AsyncFakeCloudServer:
     """ a fake server to act as the HMIP cloud"""
 
     # region __init__ & helper functions
@@ -28,22 +31,39 @@ class FakeCloudServer:
 
             self.client_auth_waiting = None  # used in auth
             self.home_id = "00000000-0000-0000-0000-000000000001"
+            self.ws = None
 
-    def __call__(self, environ, start_response):
-        request = Request(environ)
+    def __del__(self):
+        if self.ws:
+            self.ws.close()  # pragma: no cover
+
+    async def __call__(self, request):
         response = Response()
         methodname = "{}{}".format(
             request.method.lower(), request.path.replace("/", "_")
         )
-        response.content_type = "application/json;charset=UTF-8"
+        if request.method.lower() == "post":
+            request.data = await request.content.read()
+
+        response.content_type = "application/json"
         try:
-            response = self.call_method(methodname, request, response)
+            response = await self.call_method(methodname, request, response)
         except NameError as e:
             response.status_code = 404
             response.data = json.dumps({"errorCode": str(e)})
-        return response(environ, start_response)
+        except TypeError as e:
+            response = await self.call_method(methodname, request)
+        if isinstance(response, Response):  # Workaround until all methods are converted
+            res = web.Response(
+                body=(response.data),
+                status=response.status_code,
+                headers=response.headers,
+            )
+            return res
+        elif isinstance(response, web.StreamResponse):
+            return response
 
-    def call_method(self, func_name, *args, **kwargs):
+    async def call_method(self, func_name, *args, **kwargs):
         if func_name[0] == "_":
             raise NameError("Can't call internal function {}".format(func_name))
 
@@ -51,45 +71,46 @@ class FakeCloudServer:
             raise NameError("Can't find method {}".format(func_name))
 
         func = getattr(self, func_name, func_not_found)
-        return func(*args, **kwargs)  # <-- this should work!
+        return await func(*args, **kwargs)
 
     def validate_authorization(func):
-        def func_wrapper(self, request: Request, response: Response):
+        @functools.wraps(func)
+        async def func_wrapper(self, request: Request, response: Response):
             try:
                 if request.headers["CLIENTAUTH"] == self.client_auth_token:
                     for v in self.client_token_map.values():
                         if v == request.headers["AUTHTOKEN"]:
-                            return func(self, request, response)
+                            return await func(self, request, response)
             except:
                 pass
 
-            return self.errorCode(response, "INVALID_AUTHORIZATION", 403)
+            return self.errorCode("INVALID_AUTHORIZATION", 403)
 
         return func_wrapper
 
-    def errorCode(self, response: Response, message, status_code):
-        response.data = json.dumps({"errorCode": message})
-        response.status_code = status_code
-        return response
+    def errorCode(self, message, status_code):
+        return web.json_response({"errorCode": message}, status=status_code)
 
     # endregion
 
     # region home
 
     @validate_authorization
-    def post_hmip_home_getOAuthOTK(self, request: Request, response: Response):
+    async def post_hmip_home_getOAuthOTK(self, request: Request, response: Response):
         response.data = json.dumps(
             {"authToken": "C001ED", "expirationTimestamp": 1545568701680}
         )
         return response
 
     @validate_authorization
-    def post_hmip_home_getCurrentState(self, request: Request, response: Response):
+    async def post_hmip_home_getCurrentState(
+        self, request: Request, response: Response
+    ):
         response.data = json.dumps(self.data)
         return response
 
     @validate_authorization
-    def post_hmip_home_setLocation(self, request: Request, response: Response):
+    async def post_hmip_home_setLocation(self, request: Request, response: Response):
 
         js = json.loads(request.data)
 
@@ -100,12 +121,12 @@ class FakeCloudServer:
         return response
 
     @validate_authorization
-    def post_hmip_home_setPin(self, request: Request, response: Response):
+    async def post_hmip_home_setPin(self, request: Request, response: Response):
         js = json.loads(request.data)
 
         if self.pin:
             if request.headers.get("PIN", None) != str(self.pin):
-                response = self.errorCode(response, "INVALID_PIN", 403)
+                response = self.errorCode("INVALID_PIN", 403)
                 return response
 
         self.pin = js["pin"]
@@ -113,14 +134,20 @@ class FakeCloudServer:
             self.pin = None
         return response
 
+    async def post_hmip_home_getPin(self, request: Request, response: Response):
+        js = {"pin": self.pin}
+        response.data = json.dumps(js)
+        response.status_code = 200
+        return response
+
     @validate_authorization
-    def post_hmip_home_setTimezone(self, request: Request, response: Response):
+    async def post_hmip_home_setTimezone(self, request: Request, response: Response):
         js = json.loads(request.data)
         self.data["home"]["timeZoneId"] = js["timezoneId"]
         return response
 
     @validate_authorization
-    def post_hmip_home_setPowerMeterUnitPrice(
+    async def post_hmip_home_setPowerMeterUnitPrice(
         self, request: Request, response: Response
     ):
         js = json.loads(request.data)
@@ -131,7 +158,7 @@ class FakeCloudServer:
 
     # region home/security
     @validate_authorization
-    def post_hmip_home_security_setZonesActivation(
+    async def post_hmip_home_security_setZonesActivation(
         self, request: Request, response: Response
     ):
 
@@ -151,7 +178,7 @@ class FakeCloudServer:
         return response
 
     @validate_authorization
-    def post_hmip_home_security_setIntrusionAlertThroughSmokeDetectors(
+    async def post_hmip_home_security_setIntrusionAlertThroughSmokeDetectors(
         self, request: Request, response: Response
     ):
 
@@ -166,7 +193,7 @@ class FakeCloudServer:
         return response
 
     @validate_authorization
-    def post_hmip_home_security_setZoneActivationDelay(
+    async def post_hmip_home_security_setZoneActivationDelay(
         self, request: Request, response: Response
     ):
 
@@ -181,7 +208,7 @@ class FakeCloudServer:
         return response
 
     @validate_authorization
-    def post_hmip_home_security_getSecurityJournal(
+    async def post_hmip_home_security_getSecurityJournal(
         self, request: Request, response: Response
     ):
         with open(
@@ -197,7 +224,7 @@ class FakeCloudServer:
 
     # region home/heating
     @validate_authorization
-    def post_hmip_home_heating_activateAbsenceWithDuration(
+    async def post_hmip_home_heating_activateAbsenceWithDuration(
         self, request: Request, response: Response
     ):
 
@@ -213,7 +240,7 @@ class FakeCloudServer:
         return response
 
     @validate_authorization
-    def post_hmip_home_heating_activateAbsenceWithPeriod(
+    async def post_hmip_home_heating_activateAbsenceWithPeriod(
         self, request: Request, response: Response
     ):
 
@@ -227,7 +254,7 @@ class FakeCloudServer:
         return response
 
     @validate_authorization
-    def post_hmip_home_heating_deactivateAbsence(
+    async def post_hmip_home_heating_deactivateAbsence(
         self, request: Request, response: Response
     ):
 
@@ -239,7 +266,7 @@ class FakeCloudServer:
         return response
 
     @validate_authorization
-    def post_hmip_home_heating_activateVacation(
+    async def post_hmip_home_heating_activateVacation(
         self, request: Request, response: Response
     ):
 
@@ -255,7 +282,7 @@ class FakeCloudServer:
         return response
 
     @validate_authorization
-    def post_hmip_home_heating_deactivateVacation(
+    async def post_hmip_home_heating_deactivateVacation(
         self, request: Request, response: Response
     ):
 
@@ -270,7 +297,9 @@ class FakeCloudServer:
 
     # region rule
     @validate_authorization
-    def post_hmip_rule_enableSimpleRule(self, request: Request, response: Response):
+    async def post_hmip_rule_enableSimpleRule(
+        self, request: Request, response: Response
+    ):
 
         js = json.loads(request.data)
         try:
@@ -278,11 +307,11 @@ class FakeCloudServer:
             rule["active"] = js["enabled"]
             response.status_code = 200
         except:
-            response = self.errorCode(response, "INVALID_RULE", 404)
+            response = self.errorCode("INVALID_RULE", 404)
         return response
 
     @validate_authorization
-    def post_hmip_rule_setRuleLabel(self, request: Request, response: Response):
+    async def post_hmip_rule_setRuleLabel(self, request: Request, response: Response):
 
         js = json.loads(request.data)
         try:
@@ -290,7 +319,7 @@ class FakeCloudServer:
             rule["label"] = js["label"]
             response.status_code = 200
         except:
-            response = self.errorCode(response, "INVALID_RULE", 404)
+            response = self.errorCode("INVALID_RULE", 404)
         return response
 
     # endregion
@@ -298,7 +327,9 @@ class FakeCloudServer:
     # region device
 
     @validate_authorization
-    def post_hmip_device_setDeviceLabel(self, request: Request, response: Response):
+    async def post_hmip_device_setDeviceLabel(
+        self, request: Request, response: Response
+    ):
 
         js = json.loads(request.data)
         try:
@@ -306,11 +337,11 @@ class FakeCloudServer:
             d["label"] = js["label"]
             response.status_code = 200
         except:
-            response = self.errorCode(response, "INVALID_DEVICE", 404)
+            response = self.errorCode("INVALID_DEVICE", 404)
         return response
 
     @validate_authorization
-    def post_hmip_device_configuration_setRouterModuleEnabled(
+    async def post_hmip_device_configuration_setRouterModuleEnabled(
         self, request: Request, response: Response
     ):
 
@@ -323,22 +354,22 @@ class FakeCloudServer:
             ]
             response.status_code = 200
         except:
-            response = self.errorCode(response, "INVALID_DEVICE", 404)
+            response = self.errorCode("INVALID_DEVICE", 404)
         return response
 
     @validate_authorization
-    def post_hmip_device_deleteDevice(self, request: Request, response: Response):
+    async def post_hmip_device_deleteDevice(self, request: Request, response: Response):
 
         js = json.loads(request.data)
         if js["deviceId"] in self.data["devices"]:
             self.data["devices"].pop(js["deviceId"])
             response.status_code = 200
         else:
-            response = self.errorCode(response, "INVALID_DEVICE", 404)
+            response = self.errorCode("INVALID_DEVICE", 404)
         return response
 
     @validate_authorization
-    def post_hmip_device_configuration_setAcousticAlarmTiming(
+    async def post_hmip_device_configuration_setAcousticAlarmTiming(
         self, request: Request, response: Response
     ):
 
@@ -351,11 +382,11 @@ class FakeCloudServer:
             ]
             response.status_code = 200
         except:
-            response = self.errorCode(response, "INVALID_DEVICE", 404)
+            response = self.errorCode("INVALID_DEVICE", 404)
         return response
 
     @validate_authorization
-    def post_hmip_device_configuration_setAcousticAlarmSignal(
+    async def post_hmip_device_configuration_setAcousticAlarmSignal(
         self, request: Request, response: Response
     ):
 
@@ -368,11 +399,11 @@ class FakeCloudServer:
             ]
             response.status_code = 200
         except:
-            response = self.errorCode(response, "INVALID_DEVICE", 404)
+            response = self.errorCode("INVALID_DEVICE", 404)
         return response
 
     @validate_authorization
-    def post_hmip_device_configuration_setInAppWaterAlarmTrigger(
+    async def post_hmip_device_configuration_setInAppWaterAlarmTrigger(
         self, request: Request, response: Response
     ):
 
@@ -385,11 +416,11 @@ class FakeCloudServer:
             ]
             response.status_code = 200
         except:
-            response = self.errorCode(response, "INVALID_DEVICE", 404)
+            response = self.errorCode("INVALID_DEVICE", 404)
         return response
 
     @validate_authorization
-    def post_hmip_device_configuration_setAcousticWaterAlarmTrigger(
+    async def post_hmip_device_configuration_setAcousticWaterAlarmTrigger(
         self, request: Request, response: Response
     ):
 
@@ -402,11 +433,11 @@ class FakeCloudServer:
             ]
             response.status_code = 200
         except:
-            response = self.errorCode(response, "INVALID_DEVICE", 404)
+            response = self.errorCode("INVALID_DEVICE", 404)
         return response
 
     @validate_authorization
-    def post_hmip_device_configuration_setSirenWaterAlarmTrigger(
+    async def post_hmip_device_configuration_setSirenWaterAlarmTrigger(
         self, request: Request, response: Response
     ):
 
@@ -419,11 +450,11 @@ class FakeCloudServer:
             ]
             response.status_code = 200
         except:
-            response = self.errorCode(response, "INVALID_DEVICE", 404)
+            response = self.errorCode("INVALID_DEVICE", 404)
         return response
 
     @validate_authorization
-    def post_hmip_device_configuration_setClimateControlDisplay(
+    async def post_hmip_device_configuration_setClimateControlDisplay(
         self, request: Request, response: Response
     ):
 
@@ -434,11 +465,11 @@ class FakeCloudServer:
             d["functionalChannels"][channelIndex]["display"] = js["display"]
             response.status_code = 200
         except:
-            response = self.errorCode(response, "INVALID_DEVICE", 404)
+            response = self.errorCode("INVALID_DEVICE", 404)
         return response
 
     @validate_authorization
-    def post_hmip_device_configuration_setOperationLock(
+    async def post_hmip_device_configuration_setOperationLock(
         self, request: Request, response: Response
     ):
 
@@ -451,11 +482,11 @@ class FakeCloudServer:
             ]
             response.status_code = 200
         except:
-            response = self.errorCode(response, "INVALID_DEVICE", 404)
+            response = self.errorCode("INVALID_DEVICE", 404)
         return response
 
     @validate_authorization
-    def post_hmip_device_control_resetEnergyCounter(
+    async def post_hmip_device_control_resetEnergyCounter(
         self, request: Request, response: Response
     ):
 
@@ -466,11 +497,11 @@ class FakeCloudServer:
             d["functionalChannels"][channelIndex]["energyCounter"] = 0
             response.status_code = 200
         except:
-            response = self.errorCode(response, "INVALID_DEVICE", 404)
+            response = self.errorCode("INVALID_DEVICE", 404)
         return response
 
     @validate_authorization
-    def post_hmip_device_control_setSwitchState(
+    async def post_hmip_device_control_setSwitchState(
         self, request: Request, response: Response
     ):
 
@@ -481,11 +512,11 @@ class FakeCloudServer:
             d["functionalChannels"][channelIndex]["on"] = js["on"]
             response.status_code = 200
         except:
-            response = self.errorCode(response, "INVALID_DEVICE", 404)
+            response = self.errorCode("INVALID_DEVICE", 404)
         return response
 
     @validate_authorization
-    def post_hmip_device_control_setDimLevel(
+    async def post_hmip_device_control_setDimLevel(
         self, request: Request, response: Response
     ):
 
@@ -496,11 +527,11 @@ class FakeCloudServer:
             d["functionalChannels"][channelIndex]["dimLevel"] = js["dimLevel"]
             response.status_code = 200
         except:
-            response = self.errorCode(response, "INVALID_DEVICE", 404)
+            response = self.errorCode("INVALID_DEVICE", 404)
         return response
 
     @validate_authorization
-    def post_hmip_device_control_setShutterLevel(
+    async def post_hmip_device_control_setShutterLevel(
         self, request: Request, response: Response
     ):
 
@@ -512,7 +543,7 @@ class FakeCloudServer:
         return response
 
     @validate_authorization
-    def post_hmip_device_control_setSlatsLevel(
+    async def post_hmip_device_control_setSlatsLevel(
         self, request: Request, response: Response
     ):
 
@@ -525,12 +556,12 @@ class FakeCloudServer:
         return response
 
     @validate_authorization
-    def post_hmip_device_control_stop(self, request: Request, response: Response):
+    async def post_hmip_device_control_stop(self, request: Request, response: Response):
         response.status_code = 200
         return response
 
     @validate_authorization
-    def post_hmip_device_control_setSimpleRGBColorDimLevel(
+    async def post_hmip_device_control_setSimpleRGBColorDimLevel(
         self, request: Request, response: Response
     ):
 
@@ -545,7 +576,7 @@ class FakeCloudServer:
         return response
 
     @validate_authorization
-    def post_hmip_device_control_setSimpleRGBColorDimLevelWithTime(
+    async def post_hmip_device_control_setSimpleRGBColorDimLevelWithTime(
         self, request: Request, response: Response
     ):
 
@@ -561,7 +592,7 @@ class FakeCloudServer:
         return response
 
     @validate_authorization
-    def post_hmip_device_configuration_setNotificationSoundType(
+    async def post_hmip_device_configuration_setNotificationSoundType(
         self, request: Request, response: Response
     ):
 
@@ -580,7 +611,7 @@ class FakeCloudServer:
         return response
 
     @validate_authorization
-    def post_hmip_device_configuration_setAccelerationSensorEventFilterPeriod(
+    async def post_hmip_device_configuration_setAccelerationSensorEventFilterPeriod(
         self, request: Request, response: Response
     ):
 
@@ -594,7 +625,7 @@ class FakeCloudServer:
         return response
 
     @validate_authorization
-    def post_hmip_device_configuration_setAccelerationSensorTriggerAngle(
+    async def post_hmip_device_configuration_setAccelerationSensorTriggerAngle(
         self, request: Request, response: Response
     ):
 
@@ -608,7 +639,7 @@ class FakeCloudServer:
         return response
 
     @validate_authorization
-    def post_hmip_device_configuration_setAccelerationSensorSensitivity(
+    async def post_hmip_device_configuration_setAccelerationSensorSensitivity(
         self, request: Request, response: Response
     ):
 
@@ -622,7 +653,7 @@ class FakeCloudServer:
         return response
 
     @validate_authorization
-    def post_hmip_device_configuration_setAccelerationSensorNeutralPosition(
+    async def post_hmip_device_configuration_setAccelerationSensorNeutralPosition(
         self, request: Request, response: Response
     ):
 
@@ -636,7 +667,7 @@ class FakeCloudServer:
         return response
 
     @validate_authorization
-    def post_hmip_device_configuration_setAccelerationSensorMode(
+    async def post_hmip_device_configuration_setAccelerationSensorMode(
         self, request: Request, response: Response
     ):
 
@@ -654,21 +685,17 @@ class FakeCloudServer:
     # region auth
     # there is no 100% secure auth code here -> its a fake server...who cares =)
 
-    def post_hmip_auth_connectionRequest(self, request: Request, response: Response):
+    async def post_hmip_auth_connectionRequest(
+        self, request: Request, response: Response
+    ):
         if request.headers["CLIENTAUTH"] != self.client_auth_token:
-            response = self.errorCode(
-                response, "INVALID_AUTH_TOKEN", 403
-            )  # error responses must be validated against the real cloud
+            response = self.errorCode("INVALID_AUTH_TOKEN", 403)  # pragma: no cover
         elif self.client_auth_waiting is not None:
-            response = self.errorCode(
-                response, "AUTH_IN_PROCESS", 403
-            )  # error responses must be validated against the real cloud
+            response = self.errorCode("AUTH_IN_PROCESS", 403)  # pragma: no cover
         else:
             pin = request.headers.get("PIN", None)
             if pin != self.pin:
-                response = self.errorCode(
-                    response, "INVALID_PIN", 403
-                )  # error responses must be validated against the real cloud
+                response = self.errorCode("INVALID_PIN", 403)  # pragma: no cover
             else:
                 js = json.loads(request.data)
                 # TODO: add sgtin check
@@ -676,24 +703,24 @@ class FakeCloudServer:
                 response.status_code = 200
         return response
 
-    def post_hmip_auth_isRequestAcknowledged(
+    async def post_hmip_auth_isRequestAcknowledged(
         self, request: Request, response: Response
     ):
         if request.headers["CLIENTAUTH"] != self.client_auth_token:
-            response = self.errorCode(
-                response, "INVALID_AUTH_TOKEN", 403
-            )  # error responses must be validated against the real cloud
+            response = self.errorCode("INVALID_AUTH_TOKEN", 403)  # pragma: no cover
         else:
             js = json.loads(request.data)
             c_id = js["deviceId"]
             for c in self.data["clients"]:
                 if c == c_id:
-                    response = self.errorCode(response, "", 200)
+                    response = self.errorCode("", 200)
                     return response
-        response = self.errorCode(response, "INVALID_AUTH_CHALLANGE", 403)
+        response = self.errorCode("INVALID_AUTH_CHALLANGE", 403)
         return response
 
-    def post_hmip_auth_simulateBlueButton(self, request: Request, response: Response):
+    async def post_hmip_auth_simulateBlueButton(
+        self, request: Request, response: Response
+    ):
         client = {
             "homeId": self.home_id,
             "id": self.client_auth_waiting["deviceId"],
@@ -704,11 +731,11 @@ class FakeCloudServer:
         self.client_auth_waiting = None
         return response
 
-    def post_hmip_auth_requestAuthToken(self, request: Request, response: Response):
+    async def post_hmip_auth_requestAuthToken(
+        self, request: Request, response: Response
+    ):
         if request.headers["CLIENTAUTH"] != self.client_auth_token:
-            response = self.errorCode(
-                response, "INVALID_AUTH_TOKEN", 403
-            )  # error responses must be validated against the real cloud
+            response = self.errorCode("INVALID_AUTH_TOKEN", 403)  # pragma: no cover
         else:
             js = json.loads(request.data)
             c_id = js["deviceId"]
@@ -719,11 +746,11 @@ class FakeCloudServer:
 
         return response
 
-    def post_hmip_auth_confirmAuthToken(self, request: Request, response: Response):
+    async def post_hmip_auth_confirmAuthToken(
+        self, request: Request, response: Response
+    ):
         if request.headers["CLIENTAUTH"] != self.client_auth_token:
-            response = self.errorCode(
-                response, "INVALID_AUTH_TOKEN", 403
-            )  # error responses must be validated against the real cloud
+            response = self.errorCode("INVALID_AUTH_TOKEN", 403)  # pragma: no cover
         else:
             js = json.loads(request.data)
             c_id = js["deviceId"]
@@ -732,27 +759,25 @@ class FakeCloudServer:
                 response.data = json.dumps({"clientId": c_id})
                 response.status_code = 200
             else:
-                response = self.errorCode(
-                    response, "INVALID_AUTH_TOKEN", 403
-                )  # error responses must be validated against the real cloud
+                response = self.errorCode("INVALID_AUTH_TOKEN", 403)  # pragma: no cover
         return response
 
     # endregion
 
     # region Group
     @validate_authorization
-    def post_hmip_group_deleteGroup(self, request: Request, response: Response):
+    async def post_hmip_group_deleteGroup(self, request: Request, response: Response):
 
         js = json.loads(request.data)
         if js["groupId"] in self.data["groups"]:
             self.data["groups"].pop(js["groupId"])
             response.status_code = 200
         else:
-            response = self.errorCode(response, "INVALID_GROUP", 404)
+            response = self.errorCode("INVALID_GROUP", 404)
         return response
 
     @validate_authorization
-    def post_hmip_group_setGroupLabel(self, request: Request, response: Response):
+    async def post_hmip_group_setGroupLabel(self, request: Request, response: Response):
 
         js = json.loads(request.data)
         if js["groupId"] in self.data["groups"]:
@@ -760,11 +785,13 @@ class FakeCloudServer:
             g["label"] = js["label"]
             response.status_code = 200
         else:
-            response = self.errorCode(response, "INVALID_GROUP", 404)
+            response = self.errorCode("INVALID_GROUP", 404)
         return response
 
     @validate_authorization
-    def post_hmip_group_switching_setState(self, request: Request, response: Response):
+    async def post_hmip_group_switching_setState(
+        self, request: Request, response: Response
+    ):
 
         js = json.loads(request.data)
         if js["groupId"] in self.data["groups"]:
@@ -772,11 +799,11 @@ class FakeCloudServer:
             g["on"] = js["on"]
             response.status_code = 200
         else:
-            response = self.errorCode(response, "INVALID_GROUP", 404)
+            response = self.errorCode("INVALID_GROUP", 404)
         return response
 
     @validate_authorization
-    def post_hmip_group_switching_setShutterLevel(
+    async def post_hmip_group_switching_setShutterLevel(
         self, request: Request, response: Response
     ):
 
@@ -786,11 +813,11 @@ class FakeCloudServer:
             g["shutterLevel"] = js["shutterLevel"]
             response.status_code = 200
         else:
-            response = self.errorCode(response, "INVALID_GROUP", 404)
+            response = self.errorCode("INVALID_GROUP", 404)
         return response
 
     @validate_authorization
-    def post_hmip_group_heating_setSetPointTemperature(
+    async def post_hmip_group_heating_setSetPointTemperature(
         self, request: Request, response: Response
     ):
 
@@ -800,11 +827,13 @@ class FakeCloudServer:
             g["setPointTemperature"] = js["setPointTemperature"]
             response.status_code = 200
         else:
-            response = self.errorCode(response, "INVALID_GROUP", 404)
+            response = self.errorCode("INVALID_GROUP", 404)
         return response
 
     @validate_authorization
-    def post_hmip_group_heating_setBoost(self, request: Request, response: Response):
+    async def post_hmip_group_heating_setBoost(
+        self, request: Request, response: Response
+    ):
 
         js = json.loads(request.data)
         if js["groupId"] in self.data["groups"]:
@@ -812,11 +841,11 @@ class FakeCloudServer:
             g["boostMode"] = js["boost"]
             response.status_code = 200
         else:
-            response = self.errorCode(response, "INVALID_GROUP", 404)
+            response = self.errorCode("INVALID_GROUP", 404)
         return response
 
     @validate_authorization
-    def post_hmip_group_heating_setBoostDuration(
+    async def post_hmip_group_heating_setBoostDuration(
         self, request: Request, response: Response
     ):
 
@@ -826,11 +855,11 @@ class FakeCloudServer:
             g["boostDuration"] = js["boostDuration"]
             response.status_code = 200
         else:
-            response = self.errorCode(response, "INVALID_GROUP", 404)
+            response = self.errorCode("INVALID_GROUP", 404)
         return response
 
     @validate_authorization
-    def post_hmip_group_heating_setActiveProfile(
+    async def post_hmip_group_heating_setActiveProfile(
         self, request: Request, response: Response
     ):
 
@@ -840,11 +869,11 @@ class FakeCloudServer:
             g["activeProfile"] = "PROFILE_{}".format(int(js["profileIndex"]) + 1)
             response.status_code = 200
         else:
-            response = self.errorCode(response, "INVALID_GROUP", 404)
+            response = self.errorCode("INVALID_GROUP", 404)
         return response
 
     @validate_authorization
-    def post_hmip_group_heating_setControlMode(
+    async def post_hmip_group_heating_setControlMode(
         self, request: Request, response: Response
     ):
 
@@ -854,11 +883,11 @@ class FakeCloudServer:
             g["controlMode"] = js["controlMode"]
             response.status_code = 200
         else:
-            response = self.errorCode(response, "INVALID_GROUP", 404)
+            response = self.errorCode("INVALID_GROUP", 404)
         return response
 
     @validate_authorization
-    def post_hmip_group_switching_alarm_testSignalAcoustic(
+    async def post_hmip_group_switching_alarm_testSignalAcoustic(
         self, request: Request, response: Response
     ):
 
@@ -868,11 +897,11 @@ class FakeCloudServer:
             assert js["signalAcoustic"]
             response.status_code = 200
         else:
-            response = self.errorCode(response, "INVALID_GROUP", 404)
+            response = self.errorCode("INVALID_GROUP", 404)
         return response
 
     @validate_authorization
-    def post_hmip_group_switching_alarm_setSignalAcoustic(
+    async def post_hmip_group_switching_alarm_setSignalAcoustic(
         self, request: Request, response: Response
     ):
 
@@ -882,11 +911,11 @@ class FakeCloudServer:
             g["signalAcoustic"] = js["signalAcoustic"]
             response.status_code = 200
         else:
-            response = self.errorCode(response, "INVALID_GROUP", 404)
+            response = self.errorCode("INVALID_GROUP", 404)
         return response
 
     @validate_authorization
-    def post_hmip_group_switching_alarm_testSignalOptical(
+    async def post_hmip_group_switching_alarm_testSignalOptical(
         self, request: Request, response: Response
     ):
 
@@ -896,11 +925,11 @@ class FakeCloudServer:
             assert js["signalOptical"]
             response.status_code = 200
         else:
-            response = self.errorCode(response, "INVALID_GROUP", 404)
+            response = self.errorCode("INVALID_GROUP", 404)
         return response
 
     @validate_authorization
-    def post_hmip_group_switching_alarm_setSignalOptical(
+    async def post_hmip_group_switching_alarm_setSignalOptical(
         self, request: Request, response: Response
     ):
 
@@ -910,32 +939,52 @@ class FakeCloudServer:
             g["signalOptical"] = js["signalOptical"]
             response.status_code = 200
         else:
-            response = self.errorCode(response, "INVALID_GROUP", 404)
+            response = self.errorCode("INVALID_GROUP", 404)
         return response
 
     # endregion
 
-    def post_getHost(self, request: Request, response: Response):
-        data = {
-            "urlREST": self.url,
-            "urlWebSocket": self.url,  # needs to be changed
-            "apiVersion": "12",
-        }
-        response.data = json.dumps(data)
-        return response
+    async def post_getHost(self, request: Request, response: Response):
+        ws_url = self.url.replace("https", "wss")
+        data = {"urlREST": self.url, "urlWebSocket": f"{ws_url}/ws", "apiVersion": "12"}
+        return web.json_response(data)
 
     # region Fake Server settings
-    def post_hmip_fake_timeout(self, request: Request, response: Response):
+    async def post_hmip_fake_timeout(self, request: Request, response: Response):
         """this function forces a timeout on the request (2 seconds)"""
         time.sleep(2)
         response.data = json.dumps({"TIMEOUT": "TIMEOUT"})
         return response
 
-    def post_hmip_fake_loadConfig(self, request: Request, response: Response):
+    async def post_hmip_fake_loadConfig(self, request: Request, response: Response):
         js = json.loads(request.data)
         with open(Path(__file__).parent.joinpath("json_data", js["file"])) as file:
             self.data = json.load(file, encoding="UTF-8")
         return response
 
+    # endregion
 
-# endregion
+    # region websocket
+    async def get_ws(self, request):
+        self.ws = web.WebSocketResponse()
+        await self.ws.prepare(request)
+        async for msg in self.ws:
+            if msg.tp == web.WSMsgType.CLOSE:  # pragma: no cover
+                await msg.close()
+
+        return self.ws
+
+    async def post_hmip_ws_send(self, request):
+        """ this function will send specific data to the websocket """
+        await self.ws.send_json(json.loads(request.data))
+        return web.Response()
+
+    async def post_hmip_ws_sleep(self, request):
+        """ this function will send specific data to the websocket """
+        js = json.loads(request.data)
+        # the server thread must be blocked here to simulate a timeout for the client
+        # NO AWAIT VERSION!
+        time.sleep(js["seconds"])
+        return web.Response()
+
+    # endregion